+++ sample/src/main/java/net/danlew/android/joda/sample/MainActivity.java
@@ -1,0 +1,14 @@
+package net.danlew.android.joda.sample;
+
+import android.app.Activity;
+import android.os.Bundle;
+
+public class MainActivity extends Activity {
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.main);
+    }
+
+}