+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+* `#4` Added library sample app
 * `#9` Reset default DateTimeZone when TIMEZONE_CHANGED broadcast is received
 
 ## 2.3.2